# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Films API",
        "version": "0.2"
    },
    "paths": {
...
        "/api/films/{film_id}/critics": {
            "post": {
                "tags": [
                    "Critics"
                ],
                "summary": "Enregistrer une nouvelle critique",
                "description": "Store a newly created resource in storage.",
                "operationId": "eef96716ea92eb54a4faf7af89d49c22",
                "parameters": [
                    {
                        "name": "film_id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "score",
                                    "comment"
                                ],
                                "properties": {
                                    "score": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8.5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Great movie!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "Token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}